Unathi Lore Update (#1008)
# Description

This PR updates the Localizations relevant to the Reptilian species by
significantly expanding upon them as Unathi, which is how they are
typically referred to as in both /tg/station and Baystation12 derived
codebases in SS13. I have added a new "Rare" language taken from Aurora,
called Azaziba, which can only be taken by Unathi as a bonus language
trait.

Translators for this new language Azaziba are intentionally not
obtainable ingame except as spawned by Admins. This is a "Rare"
language, and shouldn't be accessible to players outside of Unathi
obtaining the trait.

The biggest change in this update however is a massive expansion upon
the Unathi Guidebook entry, providing a significant volume of Lore,
which is largely taken(with slight modifications) from
https://wiki.aurorastation.org/index.php?title=Unathi (Which, I can do
this since their wiki has a Creative Commons license). This is probably
the best way to get good quality lore for them, other than by writing it
myself.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/842d9c40-3cf3-41ab-a9fe-f0aac81c1bce)


![image](https://github.com/user-attachments/assets/f1ef8090-97a4-46f9-8548-9389606d4e3b)


![image](https://github.com/user-attachments/assets/870ab41c-fe00-45cf-8440-51cfe3f0b164)


![image](https://github.com/user-attachments/assets/09a8548a-cb98-4158-b0a8-e641078ecf80)

</p>
</details>

# Changelog

:cl:
- add: Localizations for Reptilians that acknowledge them now as Unathi.
- add: Azaziba has been added as a new "Rare" language that can only be
taken by Unathi. Translators between Azaziba and Tau-Ceti Basic do not
exist, please do not make them.
- add: SIGNIFICANTLY expanded upon the Guidebook for Unathi by including
a large volume of lore taken from the Aurora Station Wiki.