Cargo Hardsuit Market (#1024)
# Description

This PR adds several hardsuits to a new "hardsuits" category for the
cargo menu. This is part of an ongoing effort to provide reworks to
Salvage gameplay. This should pair quite well with merging Frontier's
changes to Salvage Expeditions, as well as a variety of other changes
I've been talking about. My theory is that if Expeditions are changed to
provide money rewards deposited directly to the station's account, then
Salvage can benefit the station by A: Providing station funds whether
they want to or not. And B: Be encouraged to do so because by getting
cargo a shitload of money, they will be able to directly buy the
equipment needed to go on higher tier expeditions. And to do this
without needing to gamble on wrecks for 2 hours straight, desperately
hoping you get the specific wreck that contains that security hardsuit.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/287aab37-e16a-4f59-8f8a-a1941aa2e908)

</p>
</details>

# Changelog

:cl:
- add: Logistics can now directly order a variety of hardsuits. "NT
In-Network" hardsuits being substantially cheaper thanks to backroom
deals.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>