Feat: Sleeping and Muzzle Block Vocal Emotes (#1037)
# Description
So you can actually muzzle a felinid to stop them from meowing. Or a
vulp from barking. Or a human from screaming. You get the point.

This muzzle part of this PR is a bit hardcode-y, but it's because the
muzzle simply adds a replacement accent to the entity instead of using
its own system to block speech.

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/0e5e4986-d2a0-471a-a119-579741265178

</p>
</details>

# Changelog
:cl:
- add: Sleeping and muzzled entities can no longer use vocal emotes
(scream, meow, etc).