Syndicate Listening Outpost Version 2 (#1065)
# Description

This PR does two things. First it updates the code behind the Syndicate
Listening Outpost event so that it isn't completely dogshit anymore.
Thanks to my college for teaching me Trigonometry, I now actually know
how to write a system like this from scratch. A key note is that the new
system for spawning the outpost actually accepts a List of shuttle
paths, and will randomly select a shuttle from said list to be spawned.
This allows people to potentially make "Variant" listening outposts,
such as a much larger 8-person outpost. It also now has support for
Multi-grid or map stations, via the GetSpawnableStations function, which
it uses to select a random station to spawn in reference to.

This PR also makes significant updates to the ListeningPost, extending
the outpost bridge to include new consoles, such as the Syndicate
Communications Console, which allows the LP operators to troll the
station even harder. As well as camera consoles, assuming we ever get
cross-grid cameras working... The listening post also gets a new cargo
dock, with concealed turret shutters inside the dock, and a window
facing into the cavern entrance. The cargo dock can be docked by ships,
presumably for directly exchanging supplies and equipment with the
listening post.

To make use of this dock, the Listening Post now has actual syndicate
supplies, in the form of their surplus crates, which will serve either
as loot for salvage, or supplies to be given to syndicate agents should
they have some way of getting to the outpost. Woe upon a station should
they ever team up with Nukies.

The asteroid has also had its defenses beefed up, with turrets in
various positions to guard it from spaceborne attackers. Salvage players
are REALLY going to have to work to crack the outposts open and get to
their goodies.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/67e1048a-1b5a-4495-a2f2-b0fcac90badb)


</p>
</details>

# Changelog

:cl:
- add: The Syndicate has significantly expanded their Listening Outpost
operations, featuring larger, better defended outposts. These new
outposts also serve as supply depots for Syndicate agents in the field.