Job Whitelist System (#1093)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR Port the following PRs:
https://github.com/space-wizards/space-station-14/pull/28085 and
https://github.com/DeltaV-Station/Delta-v/pull/1678

This adds a simple job whitelist system, DB Migration correctly done,
tested and working so far.
Why now? i need it.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Job WhiteList
- [x] Working with current RequirementSystem
- [x] DeltaV - Job Whitelist Panel Port

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/599b9018-aabe-4e5e-be86-dd2fd462c22c)

![image](https://github.com/user-attachments/assets/3eb5cf1e-b27d-40f7-b747-53fb298b6ff6)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Job Whitelist!