Musician Loadouts (#1108)
# Description

Fixes #1103
This PR removes the default saxophone and acoustic guitar that the
Musician spawns with. Instead the Musician now has a selection of 3 free
picks of *almost* any instrument in the game. The only instruments not
included in his list are the Supersynth, DAW, and admeme items such as
the "Musical Lung" and "Fingerless Frets". There's a few instruments
leftover for the Clown that will be given to the clown in his own
separate PR.

To make sure the musicians can also take large instruments, this PR also
takes a feature from Frontier, that of "InstrumentFlatpacks", and adds
them to the musician loadouts. So Musicians can take for instance a
Minimoog flatpack if they so desire.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/8b79e960-440f-4f7e-9b9a-bc56d30c104a)

</p>
</details>

# Changelog

:cl:
- add: Musicians can now select for free, any 3 instruments in the game
with their loadout. This now includes things like Piano flatpacks. The
only instruments not included are the Sypersynth, DAW, and "Admeme
Instruments".
- remove: Musicians no longer spawn with a Saxophone and Acoustic guitar
by default. You pick which instruments you spawn with in your loadouts
now.