Pyrokinetic Flare Power (#1138)
# Description

Behold, the counterpart to Shadeskip. Pyrokinetic Flare.
It is the "Lessermost" of all planned Pyroclastic-Anomaly Powers. This
one is quite simple. It produces a flashbang effect when cast, and
creates some glimmer. This is a natural counter to people spamming
Shadeskip all over the station, since flash effects can be used to clear
living shadows.

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/6db1db99-3985-4b6e-9766-b161394f0eee

</p>
</details>

# Changelog

:cl:
- add: Added Pyrokinetic Flare as a new Psi ability.