Remove LPO Shuttle Recalling (#1143)
# Description

The listening post being able to recall the shuttle was a little too
much for people, but luckily it and a few other things are datafields.
This PR removes the "Shuttle Recall" functionality for the
communications console on the Listening Post, while also giving a custom
broadcast announcement to the comms console so that it isn't just using
the Nuclear Operative Announcement. This way the listening post isn't
forcing people to call red alert by faking a nukie threat every round.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/97e0ca52-893b-4dd5-bc3b-bbb2a601236b)

</p>
</details>

# Changelog

:cl:
- tweak: The Syndicate Listening Post now has a custom communications
console, which no longer can recall Nanotrasen shuttles, and doesn't
sign its messages as Nuclear Operatives.