Trait Modify Factions (#955)
# Description

Made by request from Nuclear14, while also adding a trait that was
present both in Parkstation, as well as something N14 would like to
have. This PR extends the TraitSystem so that it can now also add or
remove Factions from a character. It also adds in the Animal Friend
trait from Parkstation/Fallout: New Vegas.

# Changelog

:cl:
- add: The Animal Friend trait has been added to the game. Characters
with this trait are not attacked by animals.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>