Voice Mask Refactor and Fixes (#1147)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

Based off [Voice Mask
refactor](https://github.com/space-wizards/space-station-14/pull/30798)
(The license is also different there so if need a different one, use
that PR). Had to make some changes because of the implants but overall
its basically the same!

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Fixed at least 2 bugs:
You can now use the voice mask.
You can now use the harpy voice mask implant properly.

Codebase changes:
The `VoiceMaskComponent` should not be used for masking an entities
voice. It is now used for the voice mask item!

If you were using `VoiceMaskComponent` to change an entities voice,
instead listen for the `TransformSpeakerSpeechEvent` or use the
`VoiceOverrideComponent` (If possible use the event!).

If your system was adding the `VoiceMaskComponent` when an item was
equipped and then removing it when unequipped use the
`InventoryRelayedEvent<TransformSpeakerSpeechEvent>` instead! The
`TransformSpeakerSpeechEvent` is relayed to worn clothing items now.

Also, you can and should now remove all `VoiceMaskBoundUserInterface`s
from entities that use voice masks. The interface is now stored on the
voice mask itself.

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/0835e720-f8f7-49ad-beb3-aa77d14225ec

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Beck Thompson
- fix: Fixed various bugs related to voice masks and voice mask
implants.