Prevent Listening Posts From Spawning On Planets (#1188)
# Description

Syndicate Listening Post causes servers to shit themselves if it spawns
on any map configured as a Planet, by causing the server to attempt to
spawn 7 BILLION entities. That number is neither a joke nor an
exaggeration. EntityUid is a signed 32bit integer, meaning only 2
billion entities can ever exist at a time. This would be a goddamn
performance nightmare even if I made an engine PR changing EntityUid to
a 64bit integer.

# Changelog

:cl:
- tweak: Syndicate Listening Posts can no longer spawn on a planet
surface. This isn't for balance reasons, I would love for planetside
listening posts. This is actually to address a server crashing bug
whereby listening posts on planets cause 7 billion entities to spawn.