Fix Self-Aware Crashing Client (#1247)
# Description
Quick fix to prevent the Self-Aware trait from crashing the client of a
player when examining their character's health.


[`HealthExaminableSystem.CreateMarkupSelfAware`](https://github.com/Simple-Station/Einstein-Engines/blob/6945e3027bc14eac1de0099d30f8f35b19a3034e/Content.Shared/HealthExaminable/HealthExaminableSystem.cs#L119)
attempts to access the `AnalyzableTypes` and `DetectableGroups` data
fields from
[`SelfAwareComponent`](https://github.com/Simple-Station/Einstein-Engines/blob/6945e3027bc14eac1de0099d30f8f35b19a3034e/Content.Shared/Traits/Assorted/Components/SelfAwareComponent.cs),
but since they are only initialized in the server and `null` in the
client, the client crashes. I believe this could be related to the
recent trait overhauls, but not sure. I just fixed it by adding
`AutoNetworkedField` to those fields to automatically propagate the
initialized values. Other traits might have a similar bug to this.

## Changelog

:cl: Skubman
- fix: Examining yourself with the Self-Aware trait will no longer crash
your game client.