The Great Loadout Reorganizing (#1230)
# Description

This was entirely too long overdue. I am reorganizing the entirety of
all ingame loadouts, so that it is possible to see what jobs are
actually missing loadout items. Every job in the game will have its own
loadout tab, each of them all sharing the same organization hierarchy
with *plainly visible blank spaces* that people can use to see what jobs
are in need of having things added to them.

Command category only contains Captain, Head of Personel, with space for
Centcomm roles and Blueshield if we ever add those. Instead all of the
"Department Specific Commands" are sorted with their respective
departments. Which makes it really obvious if for instance there's an
entire 2nd list of Head of Security drip...

# TODO

- [X] All of the Categories
- [x] Localize the categories
- [x] Make item groups for everything

## Organize Command

- [x] Captain
- [x] Head Of Personnel

## Organize Engineering

- [x] atmos tech
- [x] CE
- [x] senior engineer
- [x] station engineer
- [x] tech assistant

## Organize Epistemics
- [x] acolyte
- [x] cataloger
- [x] chaplain
- [x] golemancer
- [x] mystagogue
- [x] mystic
- [x] noviciate
- [x] psionic mantis

## Organize Logistics

- [x] cargo tech
- [x] courier
- [x] LO
- [x] salvage

## Organize Medical

- [x] chemist
- [x] CMO
- [x] doctor
- [x] medical intern
- [x] paramed
- [x] senior physician

## Organize Security

- [x] cadet
- [x] corpsman
- [x] detective
- [x] HOS
- [x] secoff
- [x] senior officer
- [x] warden

## Organize Service

- [x] bartender
- [x] botanist
- [x] chef
- [x] clown
- [x] janitor
- [x] lawyer
- [x] mime
- [x] musician
- [x] reporter

## GUH

- [x] Add literally everything we are missing
- [x] Do literally all of the Job specific Item Groups

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

# Changelog

:cl:
- fix: Reorganized Loadouts so that all Jobs now have their own job
specific tabs. The code for them has been thoroughly reorganized too,
such that figuring out which jobs are missing crap is way easier to do.
- add: Captain's Personal Weapon loadout category. Currently only
contains a choice between the antique laser pistol, or a pulse pistol.
Whichever choice is made will be used as a target for a traitor
objective.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>