Buff Medical Due to Shitmed Changes Cherry-Pick (#1259)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

+ Buffed applicable medication heal values.
+ Buffed applicable medication stack sizes.
+ Slightly buffed dylovene, burn, and brute chems.
+ Buffed applicable medication doAfter delay from 3s to 2s.

---


## Why / Balance
<!-- Discuss how this would affect game balance or explain why it was
changed. Link any relevant discussions or issues. -->
This change was made with the advent of the new Limb Damage system. This
PR seeks to achieve making SS14 Medical more equivalent to SS13 Medical
when it comes to healing damage values.

This was done as Bruise Packs and Ointment only heal 5 Brute or 5 Burn
respectively when applied to a limb. When each limb has a 100 Crit
threshold, this can make it extremely hard even with surgery to heal
limbs in the absence of Chemicals.

Furthermore on higher pop, this will lead to people burning through
applicable meds even faster than before Limb Damage was implemented. It
would take 34 bruise packs to heal someone with 170 Blunt spread out
across their limbs, with this change it would reduce that to 12 bruise
packs.

Full list of numerical val changes:
 - Healing component doAfter Delay reduced. was 3s, now its 2s.
 - Applicable medical items now have a max stack size of 15 (was 10).
- Ointment heals 15 of each burn type, 10 caustic (was 5 of each burn
type, 1.5 caustic)
 - Mesh now heals 20 of each burn type (was 10 of each burn type).
- Bruise Packs now heal 15 of each brute type (was 5 of each brute type)
 - Sutures now heal 20 of each brute type (was 10 of each brute type)
- Bloodpacks now heal 2.5 Bloodloss & restore 10% bloodlevel (was .5
Bloodloss and 5%)
 - Gauzes now heal 15 Slash and 20 Pierce (was 5 Slash and 10 Pierce)
 - Dylovene now heals 1.5 Poison per .5u (was 1 Poison per .5u)
 - Bicaridine now heals 2.5 Brute per .5u (Was 2 Brute per .5u)
 - Dermaline now heals 2 Burn per .5u (was 1.5 Burn per .5u)
- Epinephrine now heals 1 Burn & 1 Brute per .5u (was .5 for each per
.5u)
 - Kelotane now heals .5 Burn per .5u (was .33 Burn per .5u)
- Omnizine now heals 3 Burn, Toxin, Airloss, and Brute per .5u (was 2
per .5u)
- Lacerinol now heals 4 Slash per .5u (was 3 per .5u)
- Bruizine now heals 4 Blunt per .5u (was 3.5 per .5u) 

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Cherry-pick Medical changes merged in Goobstation

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->
---
 
 ## Breaking changes
<!-- List any breaking changes, including namespaces, public
class/method/field changes, prototype renames; and provide instructions
for fixing them.
This will be posted in #codebase-changes. -->
Should values be overtuned, they can be reduced in a future PR or before
this PR is merged.

**Changelog**
<!-- Add a Changelog entry to make players aware of new features or
changes that could affect gameplay.
Make sure to read the guidelines and take this Changelog template out of
the comment block in order for it to show up.
Changelog must have a :cl: symbol, so the bot recognizes the changes and
adds them to the game's changelog. -->
:cl:
- tweak: Increased applicable medication heal values.
- tweak: Increased stack size of applicable medications.
- tweak: Slightly increased dylovene, burn, and brute chemicals heal
values.
- tweak: Decreased Medical item application time from 3s to 2s

---------

Signed-off-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>
Signed-off-by: Tmanzxd <164098915+Tmanzxd@users.noreply.github.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>