feat: add minimum/maximum count for item groups
When there are not enough items in a group, the loadout system
will load the default items. When the selected items in a group
exceeds the maximum amount allowed, the system will just reduce
the amount of items to the max amount instead of not loading anything
at all.

TODO: Backend is fine, need to handle the client UI stuff of this