Admin Tooling Cherry Picks (#1290)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Link to every PR I cherry-picked:
1. https://github.com/new-frontiers-14/frontier-station-14/pull/2283
2. https://github.com/space-wizards/space-station-14/pull/29219
3. https://github.com/space-wizards/space-station-14/pull/30075
4. https://github.com/space-wizards/space-station-14/pull/28639
5. https://github.com/space-wizards/space-station-14/pull/32527
6. https://github.com/space-wizards/space-station-14/pull/28030
7. https://github.com/space-wizards/space-station-14/pull/28178

The main purpose of this PR is the first cherry-picked PR. It adds the
ability for admins to reply to ahelps via discord using an HTTP POST
request. See all relevant details in the initial PR by Myzumi.

---------

Co-authored-by: Myzumi <34660019+Myzumi@users.noreply.github.com>
Co-authored-by: Whatstone <166147148+whatston3@users.noreply.github.com>
Co-authored-by: Whatstone <whatston3@gmail.com>
Co-authored-by: Pieter-Jan Briers <pieterjan.briers+git@gmail.com>
Co-authored-by: to4no_fix <156101927+chavonadelal@users.noreply.github.com>
Co-authored-by: Repo <47093363+Titian3@users.noreply.github.com>
Co-authored-by: Chief-Engineer <119664036+Chief-Engineer@users.noreply.github.com>
Co-authored-by: metalgearsloth <comedian_vs_clown@hotmail.com>
Co-authored-by: Leon Friedrich <60421075+ElectroJr@users.noreply.github.com>
Co-authored-by: ShadowCommander <10494922+ShadowCommander@users.noreply.github.com>