Religious Headgear For Character Customization (#1297)
# Description

This PR adds a few articles of various kinds of religious headgear to
loadouts, namely a Hijab, Turban, and Kippah. All 3 are whitescale items
that are set as colorable, allowing players to make them whatever color
is desired.

Farida can finally have the pink Hijab she was always supposed to have.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/35fca330-98ba-439b-8e23-eec917b0d30c)

</p>
</details>

# Changelog

:cl:
- add: Added various articles of religious headgear to loadouts, such as
Hijab, Kippah, and Turban. All of these are set to allow custom colors.