Nanotrasen Representative + Blueshield Officer Jobs (#86)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Yippee

---

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Loadouts

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Nanotrasen Representative
- add: Added Blueshield Officer

---------

Co-authored-by: Aidenkrz <aiden@djkraz.com>
Co-authored-by: Icepick <122653407+Icepicked@users.noreply.github.com>
Co-authored-by: Memeji <greyalphawolf7@gmail.com>
Co-authored-by: Theapug <159912420+Teapug@users.noreply.github.com>