ChemMaster Sorting (#1310)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

[Ports this Frontier
PR.](https://github.com/new-frontiers-14/frontier-station-14/pull/2498#issuecomment-2513268858)

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: DmitriTheDemon
- tweak: The ChemMaster is now sorted.

---------

Signed-off-by: Remuchi <72476615+Remuchi@users.noreply.github.com>
Co-authored-by: Dmitri <reznoriam@gmail.com>
Co-authored-by: Whatstone <166147148+whatston3@users.noreply.github.com>
Co-authored-by: Remuchi <72476615+Remuchi@users.noreply.github.com>