Several Small Map Fixes to Europa (#1319)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Description.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->
- [x] Lack of radiator underneath the SM crystal
- [x] Wrong kind of vent and lack of connectors at the gas chambers
- [x] Inverts one of the connector pumps in the SM's closed loop
- [x] Adds a decal to this inverted pump which was missing
- [x] Lack of Mv wire under a APC on a hallway at the north side
- [x] Lack of Hv wire connection to a substation in the bridge which
powers the bridge and the bar area
- [x] Switches a medical records computer in medbay to a crew monitoring
one
- [x] Adds an autolathe to cargo
- [x] Protrudes the cargo dock which was too tight to fit the shuttle

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

Previous docking issue:

![image](https://github.com/user-attachments/assets/30ce4994-8294-40ff-8919-47fa28224dea)

Change in cargo dock design

![2024-12-06_09-25](https://github.com/user-attachments/assets/f968d153-28fe-4007-9b47-d8e36522cfc2)

Crew monitoring replacing medical records

![image](https://github.com/user-attachments/assets/719df986-fe72-4e6b-9263-9aa18b8c2237)

Autolathe added to cargo

![2024-12-06_09-19](https://github.com/user-attachments/assets/4d527179-2242-4d05-a229-d0fcc0aed360)

<h2>Atmos/SM/Wiring changes:</h2>

Added missing Hv connection to a substation in the bridge 

![image](https://github.com/user-attachments/assets/0449cf0a-83f4-4ae5-a7cf-49e2f55a5197)

Added one missing Mv wire to this APC 

![image](https://github.com/user-attachments/assets/79e939ed-1082-46e2-b572-2f5a31affa7c)

Changed regular vents into passive vents, added connector slot and
attached canisters to it

![image](https://github.com/user-attachments/assets/e8a9c686-f97e-46c3-a9e9-6b1da6776ba3)

Radiator added to SM

![2024-12-06_08-51](https://github.com/user-attachments/assets/e30549dc-1ea3-40d3-b12e-168497e9f19b)

These two pumps in the SM which were redundant, also one of them were
missing decal to the same tile

![image](https://github.com/user-attachments/assets/66e4dcdf-6b8c-4778-910e-288dcc8168ef)

Showing the changed pump after being inverted  

![image](https://github.com/user-attachments/assets/ea571ad3-c6aa-4424-ae74-11319ac0f551)

---

Recenter changes:

There were three APC in cargo and north of in which one was connected
with cargo's APC for no apparent reason and the other was missing wires
entirely. I've put the one missing wires to work in the Lv wiring around
it and cut the Lv wires were it seemed appropriate to prevent connecting
multiple APCs with Lv for no reason.


![image](https://github.com/user-attachments/assets/b7a3f1e4-5d1e-48a8-8d54-31efecb9c691)

![image](https://github.com/user-attachments/assets/669d6827-510c-475a-9039-878fbc221faf)
Excess lamps not included in the change

![image](https://github.com/user-attachments/assets/6f1bb7fc-3b5c-4f76-96da-1e362a0c3921)

![image](https://github.com/user-attachments/assets/3e25bd5b-02ff-438a-89da-91310ca3e473)

![image](https://github.com/user-attachments/assets/f2a7990d-f09f-4c44-b68b-830305743c8f)

![image](https://github.com/user-attachments/assets/953948fe-6b3b-48fb-a18f-4a1b2541ccaf)






</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Several small fixes to Europa map, namely to cargo dock, SM and
atmos piping and wires, adds autolathe to cargo and crew monitor to med