Uplink: Remove Deception Category (#1309)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

[Ports this downstream
PR.](https://github.com/Fansana/floofstation1/pull/315)

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: fenndragon
- tweak: Moved the uplink deception category into utility.

---------

Co-authored-by: fenndragon <fenndragon@gmail.com>