Trait Purchasable Mentalic Powers (#1343)
# Description

This PR adds a few of the "Mentalic" category of psionic abilities to
the Traits System, allowing them to be purchased by any character that
is also a Latent Psychic. These powers are:
- Dispel
- Metapsionic Pulse
- Xenoglossy (Also requires Natural Telepath)
- Psychognomy (Also requires Natural Telepath)

These are all quite expensive, so to help address the "Expensive trait
proliferation", I've increased the points granted by the Blindness trait
to 10. Have fun being a blind wizard. From this point on, we're going to
need to start adding more "High point value" negative traits.

# Changelog

:cl:
- add: Dispel, Metapsionic Pulse, Xenoglossy, and Psychognomy can all be
purchased during character creation using trait points.
- tweak: Blindness now grants 10 trait points instead of 6.