Port Cosmatic Drift Arrivals (#1340)
# Description

Done after a lengthy discussion with @OldDanceJacket about mapping
issues, namely the existence of "Walk-In Closets" on maps that take up
an inordinate amount of space that would otherwise be usable for other
purposes. Seriously, Saltern has a 4x6 room that contains >200 medical
uniforms, and is on a map intended only for 15 people(And maximum of 6
medbay personnel). Every map in rotation is like this.

Ostensibly, Loadouts were created to provide a solution to this problem,
by making it so that players choose their DRIP before spawning in, so
that mappers don't need to put in a whole ass room dedicated to medical
uniforms. But Mappers ended up keeping the "Walk-in Closets" even after
Loadouts were introduced to serve as a "Sanity check" in case of events
such as people not choosing their loadouts, or getting hired during the
round by the HoP to work in the medbay, etc. To solve this, me and ODJ
agreed upon adopting Cosmatic Drift's solution to this, by making it so
that the Arrivals Terminal serves as the "Walk-in Closet Hotel" that
every map can just piggyback off of. Nobody needs to make the stupid
closet rooms, on ANY map, because every map is fully guaranteed to have
every closet the players could ever need in arrivals.

# Media
<details><summary><h1>Media</h1></summary>
<p>

Dock

![image](https://github.com/user-attachments/assets/af28d549-83ab-4da0-9a4f-a5ec4bb52c7c)

Dorms(Spawn points)

![image](https://github.com/user-attachments/assets/b3d57428-ea91-4c4d-ab99-28a46e647b17)

Medbay: 

![image](https://github.com/user-attachments/assets/ebd47dce-5fb3-45fc-858d-39680d5c511a)

Security

![image](https://github.com/user-attachments/assets/0f26a112-eeff-48da-b916-15d69ede601d)

Captain/HOP

![image](https://github.com/user-attachments/assets/74ec295f-154e-4743-b36f-8161a966ee4e)

Cargo:

![image](https://github.com/user-attachments/assets/ae9f7ee1-c6c8-4c94-8dd1-13731eeb1036)

Epistemics:

![image](https://github.com/user-attachments/assets/68293436-9b69-4157-a8eb-8471c18d4f73)


</p>
</details>

# Changelog

:cl:
- add: Ported Cosmatic Drift's version of the Arrivals Terminal.
Terminal now includes every version of department clothing and material
lockers that players could ever need, allowing them to try on their
chosen job equipment and drip before departing for the station.