Traits System Anticheat (#1358)
# Description

It turns out that there was no system in place for serverside fact
checking of whether or not people have a legal traits list. Last night a
bug was reported whereby a player used Cheat Engine to give himself
every trait in the game, bypassing the points system entirely. It's not
actually possible to reduce a trait selection down to a legal list
without creating interesting race conditions, which limits my options on
how to deal with it.

So I made it a vote on the Einstein Engines discord, and the vote was
unanimous. PUNISH THE CHEATERS.

<details><summary><h1>Media</h1></summary>
<p>

https://www.youtube.com/watch?v=X2QMN0a_TrA

</p>
</details>

# Changelog


:cl:
- add: Implemented Anti-cheat for Traits. Attempting to join a round
with an illegal traits list will result in hilarious consequences.