Adjusts Cargo Market (#1377)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Long story short, cargo and command abuses the new reverse-engineered
suits and warning them as admins about it is getting quite tiring. So I
decided to put my hand to the craft and comment them out for now, until
a better way for them to be gotten is made.

I have also taken this opportunity to fix some of the longstanding
issues with cargo I have had, such as there missing sec masks in the
riot crate or there not being a way to get the swat items.

While I am still missing the properly developed versions with space
proofing and minor cold/head protection, the old non ARCS riot suit will
do for now.

This PR fixes
https://github.com/Simple-Station/Einstein-Engines/issues/1376

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Remove problematic items from cargo
- [x] Add proper swat crate

Opt:
- [x] Make proper swat set (soon:tm:)
- [x] Adjust the armor values to fit better (can be delayed)

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

I ain't got nothing chief.

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added swat suit
- add: Added swat crate
- add: Added new crate category for secure reinforced crates
- tweak: Tweaked security supplies and crates
- tweak: Tweaked the price of Cybersun's tacsuits because apparently
they weren't charging enough to properly scam salvage techs.
- tweak: Tweaked `StructuralMetallicStrong` structural protection from
10 to 80

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>