Added Argenti to the Security Loadout. (#135)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds a lethal and non-lethal Argenti for the security loadout, adds
missing magazines to the ammofab.

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added security issued Argentis with lethal and non-lethal options
into the security loadout. Rootin' tootin' cowboy shootin'!
- add: Added recipes for the magazines for the salvage rifles into the
ammo fabricator.

---------

Co-authored-by: Rosycup <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>