Assay Psionic Power (#1450)
# Description

This PR adds a new psionic power to the game, called Assay. Assay is a
more advanced information gathering tool that is available roundstart to
the Mantis. It essentially acts as the first ingame method for
characters to discover the underlying math behind Psionics, displaying
to them the target's casting stats, potentia, and metapsionic feedback
messages. These messages don't tell the caster directly which powers a
target has, instead providing hints as to what those powers might be.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/d729e373-0406-4b0b-a47c-ae7c853e72b3)


![image](https://github.com/user-attachments/assets/08457e25-56c3-4ff9-856f-ce917bd40d5a)

</p>
</details>

# Changelog

:cl:
- add: Added Assay as a new psi-power, which is available roundstart to
the Psionic Mantis. Assay allows the caster to obtain more direct
information about the statistics of a specific Psion, as well as vague
hints as to what their powers may be, if any.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: Skubman <ba.fallaria@gmail.com>
Co-authored-by: flyingkarii <123355664+flyingkarii@users.noreply.github.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>
Co-authored-by: SimpleStation Changelogs <SimpleStation14@users.noreply.github.com>