Radstation AI (#1487)
# Description

This PR adds a Station AI to Radstation! Also cleans up some of the
camera coverage. Since it's a mid to highpop map, this core is a bit
better protected with the addition of blast doors to the windows that
the AI can shut with a button. Also two cyborg spawners in the AI room,
because I know for a fact we'll want them for any future malf/traitor AI
implementation. :)

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/8a530f9a-c8d4-4469-81e9-a2d1be15bba1)

</p>
</details>

# Changelog

:cl:
- add: Station AI has been added to Radstation