Add Station AI And Ore Silo To (Almost) All Maps (#1499)
# Description

This PR adds an Ore Silo to every map in rotation. It also adds a
station AI to (almost) every map.

A decent number of the maps had to have a shitton of cameras added to
make AI actually playable. In particular Glacier is/was/probablystillis
pretty bad for AI.


<details><summary><h1>Media</h1></summary>
<p>


</p>
</details>

# Changelog

:cl:
- add: Added Station AI to (almost) every map in rotation.
- add: Added Ore Silos to every map in rotation.