Modify the Cargo Shuttle and ATS. (#1457)
# Description
Howdy Folks, I've made a few modifications to the cargo shuttle and the
automated Trading station (ATS). First, I added two more seats to the
shuttle. I also added some tools, including a wrench and a crowbar. I
also provided proper air distribution throughout the shuttle. I removed
the plastic airtight flats because the blast door opens, but can't be
closed. The same goes for the ATS. Finally, I added a station anchor, a
clean bot to the ATS, and a couple of lockers. The place doesn't get
dirty or drift away.


<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)
![Screenshot 2025-01-07
111438](https://github.com/user-attachments/assets/b5f44167-5f3c-4af7-8281-148a3d17b56d)
![Screenshot 2025-01-06
012814](https://github.com/user-attachments/assets/411f208d-8ddd-47c0-9f90-ed1dbd7b6bd0)

</p>
</details>

---

# Changelog

:cl: Mike32oz
- add: Station anchor on ATS 
- add: Clean bot and two locker
- tweak: Air distribution on cargo shuttle
- add: two seats including a wrench and a crowbar
- remove: Plastic airtight flats and replace them with tiny fans

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>