Fix Night Vision (#1554)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Fixes the bug where you could put a mouse on your head and get free
night vision. This works by adding a bool to the
SwitchableVisionOverlayComponent `IsEquipment` that will be checked for
before granting the thermal/night vision action. Port of
https://github.com/Goob-Station/Goob-Station/pull/1448

---


<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

No night vision with the mouse on your head:

![image](https://github.com/user-attachments/assets/ccde942c-7f08-405f-9215-b904e0d4dd6a)

Goggles still work:

![image](https://github.com/user-attachments/assets/a7c7e0f4-b49b-4255-86e3-511441234f02)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Piras314, Aviu
- fix: Fixed being able to get night vision from putting a mouse on your
head.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: Aviu00 <93730715+Aviu00@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>