Make Loadout Firearms Comply With Space Law (#1522)
# Description

Just because I can. I've added requirements to every firearm available
in Loadouts, and added a minimum character age requirement to them in
order to make them consistent with Space Law(Biesel Republic Law).

# Changelog

:cl:
- add: In order to comply with Biesel Republic laws governing firearm
ownership, all firearms now require that a person be 21 years of age or
older to purchase.