Change Time Requirements for Command Positions & Mystagogue (#1649)
# Description
Normalizes the time requirements for some of the command positions,
specifically the `Logistics Officer`, `Chief Engineer`, `Chief Medical
Officer`, and `Mystagogue`.

The Logistics Officer has too much overall time requirement. 40 hours
compared to other positions that only have 30 and 25? Lowered it.
The Chief Engineer and CMO have their times raised since Chief Engineer
should know more and CMO now has Surgery that they should know and be
familiar with as well.
The Mystagogue should have a slightly higher time requirement with
Psionics and should also be required to be a Mantis and Cataloguer for a
minimum of 5 hours so they know how their innate abilities and how to
deal with Psionics.

---

# Changelog
:cl:
- tweak: Logistics Officer now has 30 hours instead of 40 hours of time
requirement overall.
- tweak: Chief Engineer and CMO have 25 hours instead of 30 hours of
time requirement.
- tweak: Mystagogue now has 5 hours of time requirement as Mantis and
Cataloguer.